Extract foreground view.
