+++ Core/Core/SwiftUIViews/Pandas/GenericBackground.swift
@@ -1,0 +1,78 @@
+//
+// This file is part of Canvas.
+// Copyright (C) 2022-present  Instructure, Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as
+// published by the Free Software Foundation, either version 3 of the
+// License, or (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+//
+
+import SwiftUI
+
+struct GenericForeground: View {
+    @State private var scale = 1.0
+    @State private var dragOffset: CGSize = .zero
+    @State private var shouldTriggerDragStartFeedback = true
+    private let scene: PandaScene
+    private let feedback = UIImpactFeedbackGenerator(style: .light)
+
+    public init(scene: PandaScene) {
+        self.scene = scene
+    }
+
+    @ViewBuilder
+    public var body: some View {
+        Image(scene.foregroundFileName, bundle: .core)
+            .scaleEffect(x: scale, y: scale)
+            .gesture(pushGesture.simultaneously(with: dragGesture))
+            .offset(dragOffset)
+    }
+
+    private var pushGesture: some Gesture {
+        DragGesture(minimumDistance: 0, coordinateSpace: .local)
+            .onChanged { _ in
+                withAnimation {
+                    scale = 0.90
+                }
+            }
+            .onEnded { _ in
+                feedback.impactOccurred()
+                let animation = Animation.interpolatingSpring(stiffness: 1000, damping: 10, initialVelocity: 1)
+                withAnimation(animation) {
+                    scale = 1.0
+                }
+            }
+    }
+
+    private var dragGesture: some Gesture {
+        DragGesture(minimumDistance: 30, coordinateSpace: .global)
+            .onChanged { value in
+                var t = value.translation
+                t.width *= 0.5
+                t.height *= 0.5
+                dragOffset = t
+
+                if shouldTriggerDragStartFeedback {
+                    shouldTriggerDragStartFeedback = false
+                    feedback.impactOccurred()
+                }
+            }
+            .onEnded { _ in
+                shouldTriggerDragStartFeedback = true
+                feedback.impactOccurred()
+                let animation = Animation.interpolatingSpring(stiffness: 1000, damping: 10, initialVelocity: 1)
+                withAnimation(animation) {
+                    dragOffset = .zero
+                }
+            }
+    }
+}
